밸런스 게임 API 명세서 수정 및 dto 클래스 통합 (#369)
* refactor: 총 투표 수 필드 추가 및 인기 밸런스 게임 조회 응답 코드 작성

* chore: 밸런스 게임 투표 API 명세서 작성

* chore: 폴더를 Game에서 Bookmark로 이동

* refactor: GameRequest, Response 통합된 클래스로 관리

* refactor: VoteRequest, Response 통합된 클래스로 관리

* refactor: MemberRequest, Response 통합된 클래스로 관리

* chore: @Schema 어노테이션 dto 클래스에 추가

---------
