+++ src/main/java/balancetalk/game/presentation/GameController.java
@@ -1,16 +1,14 @@
 package balancetalk.game.presentation;
 
-import balancetalk.game.dto.GameDetailResponse;
-import balancetalk.game.dto.GameRequest;
-import balancetalk.game.dto.GameResponse;
+import static balancetalk.game.dto.GameDto.*;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.Arrays;
 import java.util.List;
 
@@ -25,8 +23,8 @@ public class GameController {
 
     @PostMapping
     @Operation(summary = "밸런스 게임 생성", description = "밸런스 게임을 생성합니다.")
-    public GameResponse createGame(@RequestBody final GameRequest request) {
-        return new GameResponse(1L, request.getTitle(), request.getOptionA(), request.getOptionB());
+    public String createGame(@RequestBody final GameRequest request) {
+        return SUCCESS_RESPONSE_MESSAGE;
     }
 
     @GetMapping("/{gameId}")
@@ -37,8 +35,8 @@ public class GameController {
 
     @PutMapping("/{gameId}")
     @Operation(summary = "밸런스 게임 수정", description = "밸런스 게임을 수정합니다.")
-    public GameResponse updateGame(@PathVariable final Long gameId, @RequestBody final GameRequest request) {
-        return new GameResponse(1L, "변경된 제목", "O", "X");
+    public String updateGame(@PathVariable final Long gameId, @RequestBody final GameRequest request) {
+        return SUCCESS_RESPONSE_MESSAGE;
     }
 
     @DeleteMapping("/{gameId}")
@@ -50,14 +48,17 @@ public class GameController {
     @GetMapping("/best")
     @Operation(summary = "인기 밸런스 게임 조회", description = "인기 있는 밸런스 게임 목록을 조회합니다.")
     public Page<GameResponse> findBestPosts(Pageable pageable) {
-        return null; // TODO 응답 객체 반환 코드 작성
+        GameResponse gameResponse1 = new GameResponse(1L, "제목1", "O", "X", 10);
+        GameResponse gameResponse2 = new GameResponse(2L, "제목2", "X", "O", 10);
+        List<GameResponse> gameResponses = Arrays.asList(gameResponse1, gameResponse2);
+        return new PageImpl<>(gameResponses, pageable, 1);
     }
 
     @GetMapping("/new")
     @Operation(summary = "새로운 밸런스 게임 조회", description = "새로 업로드 된 밸런스 게임 목록들을 조회합니다.")
     public List<GameResponse> findNewPosts() {
-        GameResponse gameResponse1 = new GameResponse(1L, "제목1", "O", "X");
-        GameResponse gameResponse2 = new GameResponse(2L, "제목2", "X", "O");
+        GameResponse gameResponse1 = new GameResponse(1L, "제목1", "O", "X", 10);
+        GameResponse gameResponse2 = new GameResponse(2L, "제목2", "X", "O", 10);
         return Arrays.asList(gameResponse1, gameResponse2);
     }
 }
